Deploy script updates
