+++ .openzeppelin/metadata-xdai.json
@@ -1,1 +1,1 @@
-{"upgradeManagerAddress":"0x7030C81EB0ADD208ecb4Fb5e0BF856C4bC51b6Ed","RewardSafeDelegateImplementationAddress":"0x56A68Bcf3c206B56610ae231EbEF9d21886e4400"}
+{"upgradeManagerAddress":"0x7030C81EB0ADD208ecb4Fb5e0BF856C4bC51b6Ed","RewardSafeDelegateImplementationAddress":"0x0b2680aA674259E0e080E109E224F91dab7d283E"}
