+++ package.json
@@ -27,7 +27,7 @@
     "deploy:sokol": "env HARDHAT_NETWORK=sokol DEPLOY_MNEMONIC=\"$SOKOL_MNEMONIC\" DEPLOY_MNEMONIC_DERIVATION_PATH=$SOKOL_MNEMONIC_DERIVATION_PATH yarn deploy",
     "deploy:sokol:skip-verify": "env SKIP_VERIFY=true yarn deploy:sokol",
     "deploy:sokol:upgrade-protocol": "env HARDHAT_NETWORK=sokol DEPLOY_MNEMONIC=\"$SOKOL_MNEMONIC\" DEPLOY_MNEMONIC_DERIVATION_PATH=$SOKOL_MNEMONIC_DERIVATION_PATH yarn deploy:upgrade-protocol",
-    "deploy:xdai": "env HARDHAT_NETWORK=xdai yarn deploy",
+    "deploy:xdai": "env HARDHAT_NETWORK=xdai DEPLOY_MNEMONIC=\"$XDAI_MNEMONIC\" DEPLOY_MNEMONIC_DERIVATION_PATH=$XDAI_MNEMONIC_DERIVATION_PATH yarn deploy",
     "deploy:xdai:skip-verify": "env SKIP_VERIFY=true yarn deploy:xdai",
     "deploy:xdai:upgrade-protocol": "env HARDHAT_NETWORK=xdai yarn deploy:upgrade-protocol",
     "deploy:localhost": "env HARDHAT_NETWORK=localhost SKIP_VERIFY=true yarn deploy && yarn deploy:localhost:upgrade-protocol",
@@ -42,8 +42,10 @@
     "deploy:status:fork:sokol": "env HARDHAT_FORKING=sokol FORK_BLOCK_NUMBER=$SOKOL_FORK_BLOCK_NUMBER DEPLOY_MNEMONIC=\"$SOKOL_MNEMONIC\" DEPLOY_MNEMONIC_DERIVATION_PATH=$SOKOL_MNEMONIC_DERIVATION_PATH yarn deploy:status:fork",
     "deploy:status:localhost": "env HARDHAT_NETWORK=localhost yarn deploy:status",
     "deploy:status:sokol": "env HARDHAT_NETWORK=sokol yarn deploy:status",
+    "deploy:status:xdai": "env HARDHAT_NETWORK=xdai yarn deploy:status",
     "deploy:proposed-diff": "env DEBUG=card-protocol.* DEPLOY_UTIL=proposed-diff ./node_modules/.bin/hardhat run ./scripts/deploy/run-util.ts",
     "deploy:proposed-diff:sokol": "env HARDHAT_NETWORK=sokol yarn deploy:proposed-diff",
+    "deploy:proposed-diff:xdai": "env HARDHAT_NETWORK=xdai yarn deploy:proposed-diff",
     "deploy:upgrade-protocol": "env DEBUG='card-protocol.*' ./node_modules/.bin/hardhat run scripts/deploy/upgrade-protocol.ts",
     "deploy:upgrade-protocol:sokol:patch": "env CARDPAY_VERSION=patch yarn deploy:sokol:upgrade-protocol",
     "deploy:upgrade-protocol:sokol:minor": "env CARDPAY_VERSION=minor yarn deploy:sokol:upgrade-protocol",
