+++ scripts/deploy/config-utils.ts
@@ -86,10 +86,7 @@ export const GNOSIS_SAFE_FACTORY =
   "0x76E2cFc1F5Fa8F6a5b3fC4c8F4788F0116861F9B";
 export const RELAY_SERVER_TX_SENDER =
   process.env.RELAY_SERVER_TX_SENDER ?? ZERO_ADDRESS;
-export const PREPAID_CARD_PROVISIONERS = [
-  RELAY_SERVER_TX_SENDER,
-  "0x3087ee149042C9c303784bb964f5C5dA28246Fe1",
-];
+export const PREPAID_CARD_PROVISIONERS = [RELAY_SERVER_TX_SENDER];
 export const GAS_FEE_RECEIVER = RELAY_SERVER_TX_SENDER;
 export const GAS_FEE_CARD_WEI = String(
   process.env.GAS_FEE_CARD_WEI ?? "1000000000000000000"
