+++ scripts/deploy/util.ts
@@ -15,7 +15,7 @@ import {
   HardhatNetworkHDAccountsConfig,
   HttpNetworkConfig,
 } from "hardhat/types";
-import { sortBy } from "lodash";
+import { compact, sortBy } from "lodash";
 import { resolve } from "path";
 import TrezorWalletProvider from "trezor-cli-wallet-provider";
 import { ZERO_ADDRESS } from "../../test/migration/util";
@@ -769,3 +769,46 @@ type SafeSignature = {
   signer: string;
   signatureBytes: string;
 };
+
+export async function guessSignatureAndDecode(txdata: string): Promise<void> {
+  try {
+    let signature = txdata.slice(0, 10);
+    let {
+      data: { results },
+      status,
+    } = await axios.get(
+      `https://www.4byte.directory/api/v1/signatures/?hex_signature=${signature}`
+    );
+
+    assert(status === 200, "api failed to return response");
+
+    let decodedCalls = compact(
+      results.map(({ text_signature }) => {
+        let iface = new Interface([`function ${text_signature}`]);
+        try {
+          return decodeEncodedCallWithInterface(iface, txdata);
+        } catch (e) {
+          console.log(e);
+          return null;
+        }
+      })
+    );
+
+    if (decodedCalls.length) {
+      console.log(
+        "Found these possible interpretations of the function call:\n\n",
+        decodedCalls.join("\n\n")
+      );
+    } else {
+      console.log(
+        "Could not find interpretation of function call from public signature db"
+      );
+    }
+
+    console.log(
+      "IMPORTANT: this relies on a public db of 4-byte signatures that are trivial to brute force collisions of. Do not trust these decodes and ensure the txdata you are signing is from a trusted source"
+    );
+  } catch (e) {
+    console.log("Failed to decode txdata, cannot show preview", e);
+  }
+}
